fix: python37 async comprehension compat
