+++ cpy_info
@@ -8,17 +8,18 @@ Filename:
 Argument count:    0
 Kw-only arguments: 0
 Number of locals:  0
-Stack size:        4
+Stack size:        5
 Flags:             NOFREE
 Constants:
-   0: <code object f at 0x00000222CEE90A50, file "", line 2>
+   0: <code object f at 0x0000028E5C631420, file "", line 2>
    1: 'f'
    2: None
 Names:
    0: f
    1: print
-   2: get_event_loop
-   3: run_until_complete
+   2: to_t
+   3: get_event_loop
+   4: run_until_complete
 
 
 
@@ -29,14 +30,15 @@ Filename:
 Argument count:    0
 Kw-only arguments: 0
 Number of locals:  0
-Stack size:        2
+Stack size:        3
 Flags:             OPTIMIZED, NEWLOCALS, NOFREE, COROUTINE
 Constants:
    0: None
-   1: <code object <genexpr> at 0x00000222CEE904B0, file "", line 3>
+   1: <code object <genexpr> at 0x0000028E5C631930, file "", line 3>
    2: 'f.<locals>.<genexpr>'
+   3: 2
 Names:
-   0: S
+   0: range
 
 
 
@@ -46,15 +48,16 @@ Name:              <genexpr>
 Filename:          
 Argument count:    1
 Kw-only arguments: 0
-Number of locals:  2
-Stack size:        9
+Number of locals:  3
+Stack size:        10
 Flags:             OPTIMIZED, NEWLOCALS, NESTED, NOFREE, ASYNC_GENERATOR
 Constants:
    0: None
-   1: 3
-   2: 5
+   1: 5
 Names:
-   0: StopAsyncIteration
+   0: S
+   1: StopAsyncIteration
 Variable names:
    0: .0
-   1: i
+   1: k
+   2: i
